This commit was generated by cvs2svn to compensate for changes in r6805, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@6806 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/c64tpi.c
@@ -93,7 +93,7 @@ _TPI_FUNC void store_pa(BYTE byte)
         ieee_is_dev = byte & 0x01;
         ieee_is_out = byte & 0x02;
 
-        parallel_cpu_set_bus(ieee_is_out ? oldpb : 0xff);
+        parallel_cpu_set_bus((BYTE)(ieee_is_out ? oldpb : 0xff));
 
         if (ieee_is_out) {
             parallel_cpu_set_ndac( 0 );
